Add files via upload
"Basketball Dribble Analysis" is a Python project that utilizes computer vision techniques to analyze basketball dribbling movements in a video. The project aims to detect and track the basketball in the video, count the number of dribbles performed by the player, and calculate the speed of the ball during dribbling. This project provides valuable insights into the performance of basketball players during dribbling drills or gameplay. The code is implemented using the OpenCV library for computer vision tasks and can be used for analyzing basketball dribbling videos in various contexts, including sports coaching, player development, and performance analysis.